# ARROW-11758: [C++][Compute] Improve summation kernel percision