--- a/cpp/src/arrow/compute/kernels/aggregate_basic_avx512.cc
+++ b/cpp/src/arrow/compute/kernels/aggregate_basic_avx512.cc
@@ -24,37 +24,11 @@ namespace aggregate {
 // ----------------------------------------------------------------------
 // Sum implementation
 
-// Round size optimized based on data type and compiler
-template <typename T>
-struct RoundSizeAvx512 {
-  static constexpr int64_t size = 32;
-};
-
-// Round size set to 64 for float/int32_t/uint32_t
-template <>
-struct RoundSizeAvx512<float> {
-  static constexpr int64_t size = 64;
-};
-
-template <>
-struct RoundSizeAvx512<int32_t> {
-  static constexpr int64_t size = 64;
-};
-
-template <>
-struct RoundSizeAvx512<uint32_t> {
-  static constexpr int64_t size = 64;
-};
+template <typename ArrowType>
+struct SumImplAvx512 : public SumImpl<ArrowType, SimdLevel::AVX512> {};
 
 template <typename ArrowType>
-struct SumImplAvx512
-    : public SumImpl<RoundSizeAvx512<typename TypeTraits<ArrowType>::CType>::size,
-                     ArrowType, SimdLevel::AVX512> {};
-
-template <typename ArrowType>
-struct MeanImplAvx512
-    : public MeanImpl<RoundSizeAvx512<typename TypeTraits<ArrowType>::CType>::size,
-                      ArrowType, SimdLevel::AVX512> {};
+struct MeanImplAvx512 : public MeanImpl<ArrowType, SimdLevel::AVX512> {};
 
 std::unique_ptr<KernelState> SumInitAvx512(KernelContext* ctx,
                                            const KernelInitArgs& args) {