@cyb70289 2021-03-05
List benchmark comparison of non-null case, which drops most.
See PR comment for performance analysis.

gcc-7.5

```
Before
------
SumKernelFloat/1048576/0        29.6 us         29.6 us        23768 bytes_per_second=33.0098G/s null_percent=0 size=1048.58k
SumKernelDouble/1048576/0       19.8 us         19.8 us        35733 bytes_per_second=49.3665G/s null_percent=0 size=1048.58k
SumKernelInt8/1048576/0          242 us          242 us         2895 bytes_per_second=4.03683G/s null_percent=0 size=1048.58k
SumKernelInt16/1048576/0        55.9 us         55.9 us        12446 bytes_per_second=17.4673G/s null_percent=0 size=1048.58k
SumKernelInt32/1048576/0        28.8 us         28.8 us        24615 bytes_per_second=33.9168G/s null_percent=0 size=1048.58k
SumKernelInt64/1048576/0        22.9 us         22.9 us        30336 bytes_per_second=42.6622G/s null_percent=0 size=1048.58k

After
-----
SumKernelFloat/1048576/0         115 us          115 us         6113 bytes_per_second=8.49307G/s null_percent=0 size=1048.58k
SumKernelDouble/1048576/0       47.5 us         47.5 us        14814 bytes_per_second=20.5647G/s null_percent=0 size=1048.58k
SumKernelInt8/1048576/0          358 us          358 us         1957 bytes_per_second=2.72924G/s null_percent=0 size=1048.58k
SumKernelInt16/1048576/0         155 us          155 us         4531 bytes_per_second=6.31986G/s null_percent=0 size=1048.58k
SumKernelInt32/1048576/0        53.9 us         53.9 us        12996 bytes_per_second=18.1266G/s null_percent=0 size=1048.58k
SumKernelInt64/1048576/0        36.2 us         36.2 us        19274 bytes_per_second=26.9443G/s null_percent=0 size=1048.58k
```


clang-9

```
Before
------
SumKernelFloat/1048576/0        28.9 us         28.9 us        24207 bytes_per_second=33.7537G/s null_percent=0 size=1048.58k
SumKernelDouble/1048576/0       24.1 us         24.1 us        29240 bytes_per_second=40.5844G/s null_percent=0 size=1048.58k
SumKernelInt8/1048576/0         60.9 us         60.9 us        11492 bytes_per_second=16.0346G/s null_percent=0 size=1048.58k
SumKernelInt16/1048576/0        35.8 us         35.8 us        19570 bytes_per_second=27.284G/s null_percent=0 size=1048.58k
SumKernelInt32/1048576/0        23.9 us         23.9 us        29181 bytes_per_second=40.8518G/s null_percent=0 size=1048.58k
SumKernelInt64/1048576/0        55.9 us         55.9 us        12516 bytes_per_second=17.4578G/s null_percent=0 size=1048.58k

After
-----
SumKernelFloat/1048576/0         139 us          139 us         5053 bytes_per_second=7.04828G/s null_percent=0 size=1048.58k
SumKernelDouble/1048576/0       46.5 us         46.5 us        14836 bytes_per_second=20.9803G/s null_percent=0 size=1048.58k
SumKernelInt8/1048576/0          322 us          322 us         2173 bytes_per_second=3.03406G/s null_percent=0 size=1048.58k
SumKernelInt16/1048576/0        71.0 us         71.0 us         9839 bytes_per_second=13.7602G/s null_percent=0 size=1048.58k
SumKernelInt32/1048576/0        40.9 us         40.9 us        17058 bytes_per_second=23.8892G/s null_percent=0 size=1048.58k
SumKernelInt64/1048576/0        45.1 us         45.1 us        15500 bytes_per_second=21.6414G/s null_percent=0 size=1048.58k
```