@cyb70289 2021-03-10
Turn this PR to draft.
Investigating a way to improve SetBitRunReader performance. Looks promising.