@jianxind 2021-03-11
> In principle I'm in favor of this. @jianxind do you want to take a look?

I don't has too much time to look the detail, sorry I switch to other internal project recently.

Anyway I agree the PR to use pairwise sum for float/double. It could be great if the int type performance can still be kept.