@cyb70289 2021-03-11
Thanks @jianxind , I do find some way to keep (actually improve) int performance for 0%, 0.01% and 1% nulls.